Move SynapseFmt and CxSynapse
To a separate synapses.py file.
