+++ nengo_loihi/synapses.py
@@ -1,0 +1,204 @@
+import warnings
+
+from nengo.utils.compat import range
+import numpy as np
+
+
+def tracing_mag_int_frac(synapses):
+    mag = synapses.tracing_mag
+    mag = mag / (synapses.size() / 100)
+
+    mag_int = int(mag)
+    # TODO: how does mag_frac actually work???
+    #  It's the x in x/128, I believe
+    mag_frac = int(128 * (mag - mag_int))
+    # mag_frac = min(int(round(1./mag_frac)), 128)
+
+    return mag_int, mag_frac
+
+
+def shift(x, s, **kwargs):
+    if s < 0:
+        return np.right_shift(x, -s, **kwargs)
+    else:
+        return np.left_shift(x, s, **kwargs)
+
+
+class SynapseFmt(object):
+    INDEX_BITS_MAP = [0, 6, 7, 8, 9, 10, 11, 12]
+    WEIGHT_BITS_MAP = [0, 1, 2, 3, 4, 5, 6, 8]
+
+    def __init__(self, wgtLimitMant=0, wgtLimitExp=0, wgtExp=0, discMaxWgt=0,
+                 learningCfg=0, tagBits=0, dlyBits=0, wgtBits=0,
+                 reuseSynData=0, numSynapses=0, cIdxOffset=0, cIdxMult=0,
+                 skipBits=0, idxBits=0, synType=0, fanoutType=0,
+                 compression=0, stdpProfile=0, ignoreDly=0):
+        self.wgtLimitMant = wgtLimitMant
+        self.wgtLimitExp = wgtLimitExp
+        self.wgtExp = wgtExp
+        self.discMaxWgt = discMaxWgt
+        self.learningCfg = learningCfg
+        self.tagBits = tagBits
+        self.dlyBits = dlyBits
+        self.wgtBits = wgtBits
+        self.reuseSynData = reuseSynData
+        self.numSynapses = numSynapses
+        self.cIdxOffset = cIdxOffset
+        self.cIdxMult = cIdxMult
+        self.skipBits = skipBits
+        self.idxBits = idxBits
+        self.synType = synType
+        self.fanoutType = fanoutType
+        self.compression = compression
+        self.stdpProfile = stdpProfile
+        self.ignoreDly = ignoreDly
+
+    @classmethod
+    def get_realWgtExp(cls, wgtExp):
+        return 6 + wgtExp
+
+    @classmethod
+    def get_scale(cls, wgtExp):
+        return 2**cls.get_realWgtExp(wgtExp)
+
+    @property
+    def realWgtExp(self):
+        return self.get_realWgtExp(self.wgtExp)
+
+    @property
+    def scale(self):
+        return self.get_scale(self.wgtExp)
+
+    @property
+    def realWgtBits(self):
+        return self.WEIGHT_BITS_MAP[self.wgtBits]
+
+    @property
+    def realIdxBits(self):
+        return self.INDEX_BITS_MAP[self.idxBits]
+
+    @property
+    def isMixed(self):
+        return self.fanoutType == 1
+
+    def bits_per_axon(self, n_weights):
+        """For an axon with n weights, compute the weight memory bits used"""
+        bits_per_weight = self.realWgtBits + self.dlyBits + self.tagBits
+        if self.compression == 0:
+            bits_per_weight += self.realIdxBits
+        elif self.compression == 3:
+            pass
+        else:
+            raise NotImplementedError("Compression %s" % (self.compression,))
+
+        SYNAPSE_FMT_IDX_BITS = 4
+        N_SYNAPSES_BITS = 6
+        bits = 0
+        synapses_per_group = self.numSynapses + 1
+        for i in range(0, n_weights, synapses_per_group):
+            n = min(n_weights - i, synapses_per_group)
+            bits_i = n*bits_per_weight + SYNAPSE_FMT_IDX_BITS + N_SYNAPSES_BITS
+            bits_i = -64 * (-bits_i // 64)
+            # ^ round up to nearest 64 (size of one int64 memory unit)
+            bits += bits_i
+
+        return bits
+
+    def set(self, **kwargs):
+        for key, value in kwargs.items():
+            assert hasattr(self, key)
+            setattr(self, key, value)
+
+    def validate(self, core=None):
+        assert -7 <= self.wgtExp <= 7
+        assert 0 <= self.tagBits < 4
+        assert 0 <= self.dlyBits < 8
+        assert 1 <= self.wgtBits < 8
+        assert 0 <= self.cIdxOffset < 16
+        assert 0 <= self.cIdxMult < 16
+        assert 0 <= self.idxBits < 8
+        assert 1 <= self.fanoutType < 4
+
+    def discretize_weights(self, w, dtype=np.int32):
+        s = 8 - self.realWgtBits + self.isMixed
+        m = 2**(8 - s) - 1
+
+        w = np.round(w / 2.**s).clip(-m, m).astype(dtype)
+        s2 = s + self.wgtExp
+        shift(w, s2, out=w)
+        np.left_shift(w, 6, out=w)
+
+        if s2 < 0:
+            warnings.warn("Lost %d extra bits in weight rounding" % (-s2,))
+
+        ws = w // self.scale
+        assert np.all(ws <= 255) and np.all(ws >= -256)
+
+        return w
+
+
+class CxSynapses(object):
+    def __init__(self, n_axons):
+        self.n_axons = n_axons
+        self.group = None
+        self.synapse_fmt = None
+        self.weights = None
+        self.indices = None
+        self.tracing = False
+        self.tracing_tau = None
+        self.tracing_mag = None
+
+    def size(self):
+        return sum(len(w) for w in self.weights)
+
+    def bits(self):
+        return sum(self.synapse_fmt.bits_per_axon(len(w))
+                   for w in self.weights)
+
+    def max_abs_weight(self):
+        return max(np.abs(w).max() if len(w) > 0 else -np.inf
+                   for w in self.weights)
+
+    def max_ind(self):
+        return max(i.max() if len(i) > 0 else -1 for i in self.indices)
+
+    def set_full_weights(self, weights):
+        self.weights = [w.astype(np.float32) for w in weights]
+        self.indices = [np.arange(w.size) for w in weights]
+        assert weights.shape[0] == self.n_axons
+
+        idxBits = int(np.ceil(np.log2(self.max_ind() + 1)))
+        assert idxBits <= SynapseFmt.INDEX_BITS_MAP[-1]
+        idxBits = next(i for i, v in enumerate(SynapseFmt.INDEX_BITS_MAP)
+                       if v >= idxBits)
+        self.format(compression=3, idxBits=idxBits, fanoutType=1,
+                    numSynapses=63, wgtBits=7)
+
+    def set_diagonal_weights(self, diag):
+        diag = diag.ravel()
+        self.weights = [d.reshape(1).astype(np.float32) for d in diag]
+        self.indices = [np.array([i]) for i in range(len(diag))]
+        assert len(self.weights) == self.n_axons
+
+        idxBits = int(np.ceil(np.log2(self.max_ind() + 1)))
+        assert idxBits <= SynapseFmt.INDEX_BITS_MAP[-1]
+        idxBits = next(i for i, v in enumerate(SynapseFmt.INDEX_BITS_MAP)
+                       if v >= idxBits)
+        self.format(compression=3, idxBits=idxBits, fanoutType=1,
+                    numSynapses=63, wgtBits=7)
+
+    def set_learning(self, tracing_tau=2, tracing_mag=1.0):
+        assert tracing_tau == int(tracing_tau), "tracing_tau must be integer"
+        self.tracing = True
+        self.tracing_tau = int(tracing_tau)
+        self.tracing_mag = tracing_mag
+        self.format(learningCfg=1, stdpProfile=0)
+        # ^ stdpProfile hard-coded for now (see loihi_interface)
+
+        mag_int, _ = tracing_mag_int_frac(self)
+        assert int(mag_int) < 2**7
+
+    def format(self, **kwargs):
+        if self.synapse_fmt is None:
+            self.synapse_fmt = SynapseFmt()
+        self.synapse_fmt.set(**kwargs)
